+++ app/apollo/apollo-core/build.gradle.kts
@@ -10,8 +10,9 @@ dependencies {
   api(libs.arrow.core)
   api(libs.coroutines.core)
   api(libs.okhttp.core)
+
+  implementation(libs.apollo.normalizedCache)
   implementation(libs.moneta)
   implementation(projects.apolloGiraffePublic)
-  implementation(libs.apollo.normalizedCache)
   implementation(projects.coreCommonPublic)
 }
