+++ app/apollo/apollo-core/src/main/kotlin/com/hedvig/android/apollo/NetworkCacheManager.kt
@@ -1,0 +1,5 @@
+package com.hedvig.android.apollo
+
+interface NetworkCacheManager {
+  fun clearCache()
+}
