+++ app/app/build.gradle.kts
@@ -114,7 +114,6 @@ dependencies {
   implementation(libs.accompanist.navigationAnimation)
   implementation(libs.accompanist.pagerIndicators)
   implementation(libs.accompanist.systemUiController)
-  implementation(libs.accompanist.webview)
   implementation(libs.adyen)
   implementation(libs.androidx.compose.animation)
   implementation(libs.androidx.compose.foundation)
@@ -177,6 +176,7 @@ dependencies {
   implementation(libs.koin.compose)
   implementation(libs.koin.workManager)
   implementation(libs.kotlin.reflect)
+  implementation(libs.kotlinx.datetime)
   implementation(libs.kotlinx.immutable.collections)
   implementation(libs.kotlinx.serialization.json)
   implementation(libs.markwon.core)
@@ -195,7 +195,6 @@ dependencies {
   implementation(projects.apolloCore)
   implementation(projects.apolloDi)
   implementation(projects.apolloGiraffePublic)
-  implementation(projects.apolloOctopusPublic)
   implementation(projects.audioPlayer)
   implementation(projects.authAndroid)
   implementation(projects.authCore)
@@ -214,12 +213,13 @@ dependencies {
   implementation(projects.featureBusinessmodel)
   implementation(projects.featureChangeaddress)
   implementation(projects.featureClaimTriaging)
+  implementation(projects.featureForever)
   implementation(projects.featureHome)
   implementation(projects.featureInsurances)
   implementation(projects.featureOdyssey)
+  implementation(projects.featureProfile)
   implementation(projects.featureTerminateInsurance)
   implementation(projects.featureTravelCertificate)
-  implementation(projects.featureForever)
   implementation(projects.hanalyticsAndroid)
   implementation(projects.hanalyticsCore)
   implementation(projects.hanalyticsFeatureFlagsPublic)
