+++ app/app/src/main/kotlin/com/hedvig/app/util/apollo/NetworkCacheManager.kt
@@ -2,13 +2,14 @@ package com.hedvig.app.util.apollo
 
 import com.apollographql.apollo3.ApolloClient
 import com.apollographql.apollo3.cache.normalized.apolloStore
+import com.hedvig.android.apollo.NetworkCacheManager
 import slimber.log.d
 import slimber.log.e
 
-class NetworkCacheManager(
+internal class NetworkCacheManagerImpl(
   private val apolloClient: ApolloClient,
-) {
-  fun clearCache() {
+) : NetworkCacheManager {
+  override fun clearCache() {
     val didClearAllRecords = apolloClient.apolloStore.clearAll()
     if (didClearAllRecords) {
       d { "Did clear entire apolloStore cache" }
