+++ app/auth/auth-core/src/main/kotlin/com/hedvig/android/auth/LogoutUseCase.kt
@@ -1,0 +1,5 @@
+package com.hedvig.android.auth
+
+interface LogoutUseCase {
+  fun invoke()
+}
