+++ app/feature/feature-forever/build.gradle.kts
@@ -12,21 +12,23 @@ android {
 }
 
 dependencies {
-  implementation(libs.coroutines.android)
-  implementation(libs.coroutines.core)
-  implementation(libs.androidx.other.dynamicAnimation)
   implementation(libs.androidx.compose.material)
   implementation(libs.androidx.compose.material3)
+  implementation(libs.androidx.compose.uiViewBinding)
   implementation(libs.androidx.lifecycle.compose)
+  implementation(libs.androidx.other.dynamicAnimation)
   implementation(libs.apollo.normalizedCache)
   implementation(libs.arrow.core)
   implementation(libs.arrow.fx)
   implementation(libs.coil.coil)
   implementation(libs.coil.compose)
+  implementation(libs.coroutines.android)
+  implementation(libs.coroutines.core)
   implementation(libs.koin.compose)
   implementation(libs.koin.core)
   implementation(libs.kotlinx.immutable.collections)
   implementation(libs.kotlinx.serialization.core)
+  implementation(libs.moneta)
   implementation(libs.slimber)
   implementation(projects.apolloCore)
   implementation(projects.apolloGiraffePublic)
@@ -39,7 +41,5 @@ dependencies {
   implementation(projects.coreUi)
   implementation(projects.languageCore)
   implementation(projects.navigationComposeTyped)
-  implementation(libs.moneta)
-  implementation(libs.androidx.compose.uiViewBinding)
   implementation(projects.navigationCore)
 }
