+++ app/feature/feature-profile/build.gradle.kts
@@ -6,15 +6,40 @@ plugins {
 }
 
 dependencies {
+  implementation(libs.accompanist.webview)
   implementation(libs.androidx.compose.material)
   implementation(libs.androidx.compose.material3.windowSizeClass)
   implementation(libs.androidx.compose.runtime)
+  implementation(libs.androidx.lifecycle.compose)
   implementation(libs.androidx.other.activityCompose)
+  implementation(libs.apollo.normalizedCache)
+  implementation(libs.apollo.runtime)
+  implementation(libs.arrow.core)
   implementation(libs.koin.android)
+  implementation(libs.koin.compose)
+  implementation(libs.koin.core)
+  implementation(libs.moneta)
+  implementation(projects.apolloCore)
+  implementation(projects.apolloGiraffePublic)
+  implementation(projects.apolloOctopusPublic)
+  implementation(projects.authCore)
+  implementation(projects.coreCommonPublic)
   implementation(projects.coreDesignSystem)
+  implementation(projects.coreIcons)
   implementation(projects.coreResources)
   implementation(projects.coreUi)
   implementation(projects.hanalyticsCore)
+  implementation(projects.hanalyticsFeatureFlagsPublic)
+  implementation(projects.languageCore)
+  implementation(projects.marketCore)
   implementation(projects.navigationComposeTyped)
   implementation(projects.navigationCore)
+  implementation(projects.theme)
+
+  testImplementation(libs.assertK)
+  testImplementation(libs.coroutines.test)
+  testImplementation(libs.junit)
+  testImplementation(libs.turbine)
+  testImplementation(projects.coreCommonTest)
+  testImplementation(projects.hanalyticsFeatureFlagsTest)
 }
